Implement --watch for localrun with cancellation support
This commit adds --watch support to localrun by using a context
to cancel and remove any running container when a change is
detected.

watchLoop has been updated to support a context.

Tested with VSCode - I suggest turning auto-save off when
using this feature.

Tested again with faas-cli up to make sure there was no
regression:

https://asciinema.org/a/594072
