+++ commands/up.go
@@ -5,20 +5,12 @@ package commands
 
 import (
 	"bufio"
+	"context"
 	"fmt"
-	"log"
 	"os"
-	"os/signal"
-	"path"
-	"path/filepath"
 	"strings"
-	"syscall"
-	"time"
 
-	"github.com/bep/debounce"
-	"github.com/fsnotify/fsnotify"
 	"github.com/go-git/go-git/v5/plumbing/format/gitignore"
-	"github.com/openfaas/faas-cli/stack"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 )
@@ -82,197 +74,17 @@ func preRunUp(cmd *cobra.Command, args []string) error {
 }
 
 func upHandler(cmd *cobra.Command, args []string) error {
-	return watchLoop(cmd, args, func(cmd *cobra.Command, args []string) error {
-		return upRunner(cmd, args)
-	})
-}
-
-func watchLoop(cmd *cobra.Command, args []string, onChange func(cmd *cobra.Command, args []string) error) error {
-
-	// Always run an initial build to freshen up
-	if err := onChange(cmd, args); err != nil {
-		return err
-	}
-
-	var services stack.Services
-	if len(yamlFile) > 0 {
-		parsedServices, err := stack.ParseYAMLFile(yamlFile, regex, filter, envsubst)
-		if err != nil {
-			return err
-		}
-
-		if parsedServices != nil {
-			services = *parsedServices
-		}
-	}
-
-	fnNames := []string{}
-	for name, _ := range services.Functions {
-		fnNames = append(fnNames, name)
-	}
-
-	fmt.Printf("[Watch] monitoring %d functions: %s\n", len(fnNames), strings.Join(fnNames, ", "))
-
 	if watch {
-		watcher, err := fsnotify.NewWatcher()
-		if err != nil {
-			return err
-		}
-		defer watcher.Close()
-
-		patterns, err := ignorePatterns()
-		if err != nil {
-			return err
-		}
-
-		matcher := gitignore.NewMatcher(patterns)
-
-		cwd, err := os.Getwd()
-		if err != nil {
-			return err
-		}
-		yamlPath := path.Join(cwd, yamlFile)
-
-		debug := os.Getenv("FAAS_DEBUG")
-		if debug == "1" {
-			fmt.Printf("[Watch] added: %s\n", yamlPath)
-		}
-		watcher.Add(yamlPath)
-
-		// map to determine which function belongs to changed files
-		// when responding to events
-		handlerMap := make(map[string]string)
-
-		for serviceName, service := range services.Functions {
-			handlerMap[serviceName] = path.Join(cwd, service.Handler)
-
-			handlerFullPath := path.Join(cwd, service.Handler)
-
-			if err := addPath(watcher, handlerFullPath); err != nil {
-				return err
-			}
-		}
-
-		signalChannel := make(chan os.Signal, 1)
-		// Exit on Ctrl+C or kill
-		signal.Notify(signalChannel, os.Interrupt, syscall.SIGTERM)
-
-		d := debounce.New(2 * time.Second)
-
-		for {
-			select {
-			case event, ok := <-watcher.Events:
-				if !ok {
-					return fmt.Errorf("watcher's Events channel is closed")
-				}
-				log.Printf("%s %s", event.Op, event.Name)
-
-				if strings.HasSuffix(event.Name, ".swp") || strings.HasSuffix(event.Name, "~") || strings.HasSuffix(event.Name, ".swx") {
-					continue
-				}
-
-				if event.Op == fsnotify.Write || event.Op == fsnotify.Create || event.Op == fsnotify.Remove || event.Op == fsnotify.Rename {
-
-					info, err := os.Stat(event.Name)
-					if err != nil {
-						continue
-					}
-					ignore := false
-					if matcher.Match(strings.Split(event.Name, "/"), info.IsDir()) {
-						ignore = true
-					}
-
-					// exact match first
-					target := ""
-					for fnName, fnPath := range handlerMap {
-						if event.Name == fnPath {
-							target = fnName
-						}
-					}
-
-					// fuzzy match after, if none matched exactly
-					if target == "" {
-						for fnName, fnPath := range handlerMap {
-							log.Printf("Checking %s against %s", event.Name, fnPath)
-
-							if strings.HasPrefix(event.Name, fnPath) {
-								target = fnName
-							}
-						}
-					}
-
-					// New sub-directory added for a function, start tracking it
-					if event.Op == fsnotify.Create && info.IsDir() && target != "" {
-						if err := addPath(watcher, event.Name); err != nil {
-							return err
-						}
-					}
-
-					if !ignore {
-						if target == "" {
-							fmt.Printf("Rebuilding %d functions reason: %s to %s\n", len(fnNames), event.Op, event.Name)
-						} else {
-							fmt.Printf("Reloading %s reason: %s to %s\n", target, event.Op, event.Name)
-						}
-					}
-
-					if !ignore {
-						d(func() {
-							filter = target
-
-							go func() {
-								// Assign --filter to "" for all functions if we can't determine the
-								// changed function to direct the calls to build/push/deploy
-
-								if err := onChange(cmd, args); err != nil {
-									fmt.Println("Error rebuilding: ", err)
-									os.Exit(1)
-								}
-							}()
-						})
-					}
-
-				}
-
-			case err, ok := <-watcher.Errors:
-				if !ok {
-					return fmt.Errorf("watcher's Errors channel is closed")
-				}
-				return err
-
-			case <-signalChannel:
-				watcher.Close()
-				return nil
-			}
-		}
+		return watchLoop(cmd, args, func(cmd *cobra.Command, args []string, ctx context.Context) error {
+			return upRunner(cmd, args, ctx)
+		})
 	}
-	return nil
+
+	ctx := context.Background()
+	return upRunner(cmd, args, ctx)
 }
 
-func addPath(watcher *fsnotify.Watcher, rootPath string) error {
-	debug := os.Getenv("FAAS_DEBUG")
-
-	return filepath.WalkDir(rootPath, func(subPath string, d os.DirEntry, err error) error {
-		if err != nil {
-			return err
-		}
-
-		if d.IsDir() {
-			if err := watcher.Add(subPath); err != nil {
-				return fmt.Errorf("unable to watch %s: %s", subPath, err)
-			}
-
-			if debug == "1" {
-				fmt.Printf("[Watch] added: %s\n", subPath)
-			}
-		}
-
-		return nil
-	})
-
-}
-
-func upRunner(cmd *cobra.Command, args []string) error {
+func upRunner(cmd *cobra.Command, args []string, ctx context.Context) error {
 	if err := runBuild(cmd, args); err != nil {
 		return err
 	}
